comenzando internacionalizacion

+++ courier/db/migrate/20110422230512_create_paquetes.rb
@@ -2,7 +2,7 @@ class CreatePaquetes < ActiveRecord::Migration
   def self.up
     create_table :paquetes do |t|
       t.string :nombre
-      t.integer :peso
+      t.float :peso
       t.string :descripcion
       t.references :ordens
       t.references :personas
